All board-scoring gene parameters are interpolated
Namely, the Castling Possible Gene and the Sphere of Influence Gene have
parameters that might be different at the beginning and end of the game.
The Look Ahead Gene is not interpolated since its parameters determine
the estimates for game progress which control interpolation. The Total
Force Gene is not interpolated because (other than seeming awkward with
so many more parameters) is that the gene is referenced by other
interpolated genes and too much interolation of interpolated data seems
like it would make evolutionary pressure weaker (yeah, that sounds
good).