First attempt at Draw Value Gene
See comments in issue #82. Basically, since a draw is the same value for
a player and an opponent, the minimax algorithm doesn't handle it
correctly. The value should not switch sign when moving between levels
of the minimax game tree.

More work needed, possibly adding a special bool is_draw member to the
Game_Tree_Node_Result class.