Revert "Add -Wconversion to clang; fix errors found"
This reverts commit 56f3f9679267d22592376a3dae4327d68d412640.

When compiled with clang, the perft tests show no difference after the
changes in this commit.

When compiled with gcc, the changes in this commit cause a significant
slowdown in the perft tests. Given this and the noisy type-casting that
was added, I'm reverting this commit. Maybe another strategy in the
future will work better.