Only respond to `ping` when thinking is done
When ping is used for synchronizing purposes, the pong response is only
expected after every other command has been processed by the engine.
So, since the method `setup_turn()` is entered when the engine
is done with everything and waiting for more commands, that is the
proper place for responding to ping.