Fix en passant hash problem
Using the new identical_boards_have_identical_hashes() test,
a problem was found with the en passant hash calculations.
It was found that, when a double pawn move did not result
in a legal en passant move for the other side, the en passant
hash value was XORed twice: once in the fix_en_passant_hash()
method and once in update_board().

The current strategy has the en passant target cleared if there is
no legal en passant move, but also saving that useless en passant
target for writing the FEN string.

The new test has also revealed a problem with castling hashes, which
is currently being investigated.