Disregard move count of player that is not moving
The number of moves that a player had prior to making a move is probably
not relevant to the evalution of the current position. So, return zero
for the number of moves of the player not on move. This may also serve
to take into account the tempo of the player on move.