Fix bug in printing commentary due to new gene
The Minimax_AI::reset_search_stats() method needs to be called before
every move search to make sure the commentary lists line up with the
game move lists. When the new gene was added with a new way of choosing
moves, that method call was skipped, which led to prior game commentary
being used for newly completed games. This lead to garbage in the
variations written to PGN files since the moves were illegal and not
actually played.