Stop iterative deepening when endgame is found
If a win is the result at depth N, then there's no point in searching at
depth N+1. If a loss is the result at depth N, then that means every
move on the current board leads to a loss at depth N or sooner.

In either case, don't search deeper.