Replace begin()/end() pairs with ranges algorithms
The comparison and equality operators in Minimax_AI were needed to
work with the ranges::sort algorithm that requires all six comparisons
(==, !=, <, <=, >, >=) to be defined.

Also, in an algorithm which uses a lambda expression like [](auto x) { return x; }
for a parameter, that parameter can be replaced with std::identity() from
<functional>.

Finally, the pseudo-containers for squares don't work with ranges, probably
because of some deficiency in the various Square iterators. More investigation
is needed.

Partial fix to #109.