Automate spacing after options
If an option doesn't have a description, then another related option
will be written next to allow grouping similar options into one description.

In the case where there is not description, such as the "arg param" examples
at the top, an extra newline is needed before the next bit of text.

This avoids the need for a help.add_option("") line purely for typesetting
purposes.