Search further when king is in check
The number of moves that lead to check and the moves out of check
are usually both small, so the search should always continue when the
king is in check. Check is not a stable position to examine.

Now, the king's value is the Piece Strength Gene is completely
meaningless. The kings are always on the board, so the Total Strength
Gene cancels them out. The kings are never in check when evaluating,
so they are never attacked when the Opponent  Pieces Targeted gene
runs.