Correct board hash collisions
Added test that proves that two boards with the same pieces on the same
squares but with different legal en passant captures have the same hash.
Corrected hash function to give different hashes to boards with different
legal moves.

Only 8 hashes are required since the 16 possible en passant targets are
identified with the file of the target and the color of the player to move.