Plot order priority of Move Sorters
Plot a moving average of the order in which the various move sorters
are arranged, using a long-window moving average since there are only
a few values and they change discontinuously. The slower, manual version
of the moving average implemented here is because the built-in function
takes up all the available memory and eventually hangs the computer
which has to spend all of its time swapping to disk.