Remove Board::blocked_attacks structure
This structure that kept track of which squares would be attacked if
blocking pieces were removed was not being updated correctly, nor
does there seem to be a way to update it correctly without significant
slowdown. The problem was discovered by trying to use the blocking
structure to check for pinned pieces.

Consider the following portion of a board showing ranks 1 and 2:
```
 . . . . . . . .
 K . . P . r r .
```
There is a blocked attack on the king due to the black rook on the f-file
and the white pawn on the d-file. If one of the rooks moved to a different
rank, like this
```
 . . . . . . r .
 K . . P . r . .
```
Then the blocked moves would be removed from the entire row, indicating
that there was no longer a blocked attack on the white king. Fixing this
would involve checking each piece along the way as blocked attacks were
removed to see if another piece would still be supplying the blocked attack.

Another possible fix is to keep a count of blocked attacks, but that is a much
larger structure (std::array<int> vs std::bitset<16>), so would probably be
much more expensive to copy. This will be left to future experimentation.