Do not exit game loop before outsider says so
When using the UCI protocol, the previous code would effectively
ignore the `go` command and exit the game loop when the game
was over from a rule. This could cause Genetic_Chess to hang
while waiting for the next command that never comes.

The fix here is to return immediately when the result of a move
ends the game. This way, reading stdin resumes only after
reentering setup_turn() and no commands are missed.