Activate/deactivate genes based on game progress
Instead of a single Enabled parameter, let the genome evolve
the two points at which the gene is activated and then
deactivated. Any single interval during which the gene is active
can be evolved. It is interesting to see which genes are turned
on for the opening, middlegame, and endgame. It's even
possible for genes to be completely deactivated by setting
the activation and deactivation points to be the same.

Two exceptions: (1) The Piece Strength Gene does not have
activation parameters because this gene is needed for
calculating game progress. (2) The Look Ahead Gene does
not have activation parameters because deactivating this
gene results in immediate loss, wasting time and mutations.

Also, update plots to put both activation values on the same
plot.