Mutations can disable genes (#141)
* Genes can be disabled

With small probability (1/10,000 right now) a gene can be
completely disabled. This is a tradeoff between less
information and more time.

Relevant issue: #126

* Use less contrasty color pattern in opening plot

* Fix typing

* Remove outdated gene

* Use words instead of numbers for Enabled

* Make gene enable/disable probability configurable

* Improve gene enabled plots

- Increase noise to spread out points more
- Draw line between enabled and disabled points

* Activate/deactivate genes based on game progress

Instead of a single Enabled parameter, let the genome evolve
the two points at which the gene is activated and then
deactivated. Any single interval during which the gene is active
can be evolved. It is interesting to see which genes are turned
on for the opening, middlegame, and endgame. It's even
possible for genes to be completely deactivated by setting
the activation and deactivation points to be the same.

Two exceptions: (1) The Piece Strength Gene does not have
activation parameters because this gene is needed for
calculating game progress. (2) The Look Ahead Gene does
not have activation parameters because deactivating this
gene results in immediate loss, wasting time and mutations.

Also, update plots to put both activation values on the same
plot.

* Delete unused function

Add more information to success_probability() test failure.

* Undo cosmetic change

* Simplify check for Gene priorities or activations

* Allow activation points to be reversed ordered

Similarly to allowing the values to range from -0.05 or 1.05,
this will allow mutations to deactivate a gene with some
margin for error so that genes don't reactivate for one move.

* Plot when genes are inactive

* Consolidate plot options

* Create method for mutating Gene activations

* Fix variable name

* Simplify gene activation plotting