All rules implemented, but very slow
First obvious source of slowness: Moves are stored in Pieces upon construction, including pieces that are only used for reference. For example, Board::piece_is_pinned_to_king() constructs a pawn and a king, which now generates the moves generators. This will probably require moving the Move generators outside the class so that the Piece classes only contain references to them.