Fix dangling reference in C++20
According to GCC, the range in the for loop (container | std::views::drop(1)) is a temporary that is destroyed at the end of the for-loop declaration. This means that piece is a dangling reference. In order to get gcc to compile with all warnings as errors, the range is created outside the loop. This commit can be undone in the C++23 branch due to [1] (less technical explanation [2]).

[1] https://www.open-std.org/jtc1/sc22/wg21/docs/papers/2022/p2718r0.html
[2] https://www.cppstories.com/2024/cpp23_lang/#extend-lifetime-of-temporaries-in-range-based-for